+++ packages/nuxt/__tests__/nuxt.spec.ts
@@ -1,0 +1,32 @@
+/**
+ * @vitest-env node
+ */
+import { fileURLToPath } from 'node:url'
+import { describe, it, expect } from 'vitest'
+import { setup, $fetch } from '@nuxt/test-utils'
+
+await setup({
+  server: true,
+  rootDir: fileURLToPath(new URL('../playground', import.meta.url)),
+  nuxtConfig: {
+    hooks: {
+      'vite:extendConfig'(config, { isClient }) {
+        config.define!.__BROWSER__ = isClient
+      },
+    },
+    vite: {
+      define: {
+        __DEV__: false,
+        __TEST__: true,
+        __FEATURE_PROD_DEVTOOLS__: false,
+      },
+    },
+  },
+})
+
+describe('works with nuxt', async () => {
+  it('works on ssr', async () => {
+    const html = await $fetch('/')
+    expect(html).toContain('Count: 101')
+  })
+})