+++ packages/nuxt/shims.d.ts
@@ -1,5 +1,5 @@
 declare namespace NodeJS {
   export interface Process {
-      server: boolean
+    server: boolean
   }
 }