+++ packages/nuxt/src/module.ts
@@ -5,9 +5,12 @@ import {
   defineNuxtModule,
   addPlugin,
   isNuxt2,
-  addAutoImport,
+  addImports,
   createResolver,
+  resolveModule,
+  addImportsDir,
 } from '@nuxt/kit'
+import type { NuxtModule } from '@nuxt/schema'
 
 export interface ModuleOptions {
   /**
@@ -19,40 +22,40 @@ export interface ModuleOptions {
   disableVuex?: boolean
 
   /**
-   * Array of auto imports to be added to the nuxt.config.js file.
+   * Automatically add stores dirs to the auto imports. This is the same as
+   * directly adding the dirs to the `imports.dirs` option. If you want to
+   * also import nested stores, you can use the glob pattern `./stores/**`
    *
-   * @example
-   * ```js
-   * autoImports: [
-   *  // automatically import `defineStore`
-   *  'defineStore',
-   *  // automatically import `defineStore` as `definePiniaStore`
-   *  ['defineStore', 'definePiniaStore',
-   * ]
-   * ```
-   *
+   * @default `['./stores']`
    */
-  autoImports?: Array<string | [string, string]>
+  storesDirs?: string[]
 }
 
-export default defineNuxtModule<ModuleOptions>({
+const module: NuxtModule<ModuleOptions> = defineNuxtModule<ModuleOptions>({
   meta: {
     name: 'pinia',
     configKey: 'pinia',
     compatibility: {
-      nuxt: '^2.0.0 || ^3.0.0',
+      nuxt: '^2.0.0 || ^3.0.0-rc.5',
       bridge: true,
     },
   },
   defaults: {
     disableVuex: true,
-    autoImports: [],
+    storesDirs: ['./stores'],
   },
   setup(options, nuxt) {
     const resolver = createResolver(import.meta.url)
 
     // Disable default Vuex store (Nuxt v2.10+ only)
-    if (nuxt.options.features && options.disableVuex && isNuxt2()) {
+    if (
+      // @ts-expect-error: no feature flag anymore or private?
+      nuxt.options.features &&
+      // ts
+      options.disableVuex &&
+      isNuxt2()
+    ) {
+      // @ts-expect-error: same
       nuxt.options.features.store = false
     }
 
@@ -60,24 +63,41 @@ export default defineNuxtModule<ModuleOptions>({
     nuxt.options.build.transpile.push(resolver.resolve('./runtime'))
 
     // Make sure we use the mjs build for pinia
-    nuxt.options.alias.pinia = 'pinia/dist/pinia.mjs'
+    nuxt.options.alias.pinia =
+      nuxt.options.alias.pinia ||
+      // FIXME: remove this deprecated call. Ensure it works in Nuxt 2 to 3
+      resolveModule('pinia/dist/pinia.mjs', {
+        paths: [nuxt.options.rootDir, import.meta.url],
+      })
 
-    // Add runtime plugin
-    if (isNuxt2()) {
-      addPlugin(resolver.resolve('./runtime/plugin.vue2'))
-    } else {
-      addPlugin(resolver.resolve('./runtime/plugin.vue3'))
-    }
+    nuxt.hook('prepare:types', ({ references }) => {
+      references.push({ types: '@pinia/nuxt' })
+    })
+
+    // Add runtime plugin before the router plugin
+    // https://github.com/nuxt/framework/issues/9130
+    nuxt.hook('modules:done', () => {
+      if (isNuxt2()) {
+        addPlugin(resolver.resolve('./runtime/plugin.vue2'))
+      } else {
+        addPlugin(resolver.resolve('./runtime/plugin.vue3'))
+      }
+    })
 
     // Add auto imports
     const composables = resolver.resolve('./runtime/composables')
-    addAutoImport([
+    addImports([
+      { from: composables, name: 'defineStore' },
+      { from: composables, name: 'acceptHMRUpdate' },
       { from: composables, name: 'usePinia' },
-      ...options.autoImports.map((imports) =>
-        typeof imports === 'string'
-          ? { from: composables, name: imports }
-          : { from: composables, name: imports[0], as: imports[1] }
-      ),
     ])
+
+    if (options.storesDirs) {
+      for (const storeDir of options.storesDirs) {
+        addImportsDir(resolver.resolve(nuxt.options.rootDir, storeDir))
+      }
+    }
   },
 })
+
+export default module