+++ packages/nuxt/src/runtime/plugin.vue2.ts
@@ -5,13 +5,12 @@ import { createPinia, setActivePinia, PiniaVuePlugin } from 'pinia'
 const Vue = 'default' in _Vue2 ? (_Vue2 as any).default : _Vue2
 Vue.use(PiniaVuePlugin)
 
-export default (context: any, inject: any) => {
+export default (context: any, provide: any) => {
   const pinia = createPinia()
   context.app.pinia = pinia
   setActivePinia(pinia)
 
   // add access to `$nuxt`
-  // @ts-expect-error: _p is internal
   pinia._p.push(({ store }) => {
     // make it non enumerable so it avoids any serialization and devtools
     Object.defineProperty(store, '$nuxt', { value: context })
@@ -26,7 +25,7 @@ export default (context: any, inject: any) => {
   }
 
   // Inject $pinia
-  inject('pinia', pinia)
+  provide('pinia', pinia)
 }
 
 declare module 'pinia' {