+++ packages/pinia/__tests__/store.spec.ts
@@ -1,7 +1,8 @@
+import { beforeEach, describe, it, expect, vi } from 'vitest'
 import { createPinia, defineStore, setActivePinia } from '../src'
 import { mount } from '@vue/test-utils'
 import { defineComponent, getCurrentInstance, nextTick, watch } from 'vue'
-import { mockWarn } from 'jest-mock-warn'
+import { mockWarn } from './vitest-mock-warn'
 
 describe('Store', () => {
   mockWarn()
@@ -59,13 +60,13 @@ describe('Store', () => {
       id: 'main',
       state: () => ({ n: 0 }),
     })
-    const TestComponent = {
+    const TestComponent = defineComponent({
       template: `<div>{{ store. n }}</div>`,
       setup() {
         const store = useStore()
         return { store }
       },
-    }
+    })
     const w1 = mount(TestComponent, { global: { plugins: [pinia] } })
     const w2 = mount(TestComponent, { global: { plugins: [pinia] } })
     expect(w1.text()).toBe('0')
@@ -80,7 +81,7 @@ describe('Store', () => {
   it('can be reset', () => {
     const store = useStore()
     store.$state.a = false
-    const spy = jest.fn()
+    const spy = vi.fn()
     store.$subscribe(spy, { flush: 'sync' })
     expect(spy).not.toHaveBeenCalled()
     store.$reset()
@@ -139,7 +140,7 @@ describe('Store', () => {
 
   it('can replace its state', () => {
     const store = useStore()
-    const spy = jest.fn()
+    const spy = vi.fn()
     watch(() => store.a, spy, { flush: 'sync' })
     expect(store.a).toBe(true)
 
@@ -202,7 +203,7 @@ describe('Store', () => {
 
     const store = useStore(pinia)
 
-    const spy = jest.fn()
+    const spy = vi.fn()
     watch(() => store.n, spy)
 
     expect(spy).toHaveBeenCalledTimes(0)
@@ -319,7 +320,7 @@ describe('Store', () => {
     })
 
     const store = useStore(pinia)
-    const spy = jest.fn()
+    const spy = vi.fn()
 
     store.$subscribe(spy, { flush: 'sync' })
     pinia.state.value.main.n++
@@ -377,5 +378,11 @@ describe('Store', () => {
     expect(
       `[🍍]: A getter cannot have the same name as another state property. Rename one of them. Found with "anyName" in store "main".`
     ).toHaveBeenWarnedTimes(1)
+  })
+
+  it('throws an error if no store id is provided', () => {
+    expect(() => defineStore({} as any)).toThrowError(
+      /must be passed a store id/
+    )
   })
 })