+++ packages/pinia/src/devtools/plugin.ts
@@ -35,6 +35,7 @@ const componentStateTypes: string[] = []
 
 const MUTATIONS_LAYER_ID = 'pinia:mutations'
 const INSPECTOR_ID = 'pinia'
+const { assign } = Object
 
 /**
  * Gets the displayed name of a store in devtools
@@ -114,6 +115,29 @@ export function registerPiniaDevtools(app: DevtoolsApp, pinia: Pinia) {
             tooltip: 'Import the state from a JSON file',
           },
         ],
+        nodeActions: [
+          {
+            icon: 'restore',
+            tooltip: 'Reset the state (with "$reset")',
+            action: (nodeId) => {
+              const store = pinia._s.get(nodeId)
+              if (!store) {
+                toastMessage(
+                  `Cannot reset "${nodeId}" store because it wasn't found.`,
+                  'warn'
+                )
+              } else if (typeof store.$reset !== 'function') {
+                toastMessage(
+                  `Cannot reset "${nodeId}" store because it doesn't have a "$reset" method implemented.`,
+                  'warn'
+                )
+              } else {
+                store.$reset()
+                toastMessage(`Store "${nodeId}" reset.`)
+              }
+            },
+          },
+        ],
       })
 
       api.on.inspectComponent((payload, ctx) => {
@@ -134,7 +158,7 @@ export function registerPiniaDevtools(app: DevtoolsApp, pinia: Pinia) {
               value: store._isOptionsAPI
                 ? {
                     _custom: {
-                      value: store.$state,
+                      value: toRaw(store.$state),
                       actions: [
                         {
                           icon: 'restore',
@@ -144,7 +168,11 @@ export function registerPiniaDevtools(app: DevtoolsApp, pinia: Pinia) {
                       ],
                     },
                   }
-                : store.$state,
+                : // NOTE: workaround to unwrap transferred refs
+                  Object.keys(store.$state).reduce((state, key) => {
+                    state[key] = store.$state[key]
+                    return state
+                  }, {} as StateTree),
             })
 
             if (store._getters && store._getters.length) {
@@ -358,7 +386,7 @@ function addStoreToDevtools(app: DevtoolsApp, store: StoreGeneric) {
 
       store._customProperties.forEach((name) => {
         watch(
-          () => unref(store[name]),
+          () => unref<unknown>(store[name]),
           (newValue, oldValue) => {
             api.notifyComponentUpdate()
             api.sendInspectorState(INSPECTOR_ID)
@@ -393,15 +421,12 @@ function addStoreToDevtools(app: DevtoolsApp, store: StoreGeneric) {
           const eventData: TimelineEvent = {
             time: now(),
             title: formatMutationType(type),
-            data: {
-              store: formatDisplay(store.$id),
-              ...formatEventData(events),
-            },
+            data: assign(
+              { store: formatDisplay(store.$id) },
+              formatEventData(events)
+            ),
             groupId: activeAction,
           }
-
-          // reset for the next mutation
-          activeAction = undefined
 
           if (type === MutationType.patchFunction) {
             eventData.subtitle = '⤵️'
@@ -482,7 +507,11 @@ let activeAction: number | undefined
  * @param store - store to patch
  * @param actionNames - list of actionst to patch
  */
-function patchActionForGrouping(store: StoreGeneric, actionNames: string[]) {
+function patchActionForGrouping(
+  store: StoreGeneric,
+  actionNames: string[],
+  wrapWithProxy: boolean
+) {
   // original actions of the store as they are given by pinia. We are going to override them
   const actions = actionNames.reduce((storeActions, actionName) => {
     // use toRaw to avoid tracking #541
@@ -492,23 +521,30 @@ function patchActionForGrouping(store: StoreGeneric, actionNames: string[]) {
 
   for (const actionName in actions) {
     store[actionName] = function () {
-      // setActivePinia(store._p)
       // the running action id is incremented in a before action hook
       const _actionId = runningActionId
-      const trackedStore = new Proxy(store, {
-        get(...args) {
-          activeAction = _actionId
-          return Reflect.get(...args)
-        },
-        set(...args) {
-          activeAction = _actionId
-          return Reflect.set(...args)
-        },
-      })
-      return actions[actionName].apply(
+      const trackedStore = wrapWithProxy
+        ? new Proxy(store, {
+            get(...args) {
+              activeAction = _actionId
+              return Reflect.get(...args)
+            },
+            set(...args) {
+              activeAction = _actionId
+              return Reflect.set(...args)
+            },
+          })
+        : store
+
+      // For Setup Stores we need https://github.com/tc39/proposal-async-context
+      activeAction = _actionId
+      const retValue = actions[actionName].apply(
         trackedStore,
         arguments as unknown as any[]
       )
+      // this is safer as async actions in Setup Stores would associate mutations done outside of the action
+      activeAction = undefined
+      return retValue
     }
   }
 }
@@ -519,8 +555,8 @@ function patchActionForGrouping(store: StoreGeneric, actionNames: string[]) {
 export function devtoolsPlugin<
   Id extends string = string,
   S extends StateTree = StateTree,
-  G /* extends GettersTree<S> */ = _GettersTree<S>,
-  A /* extends ActionsTree */ = _ActionsTree
+  G extends object = _GettersTree<S>,
+  A extends object = _ActionsTree
 >({ app, store, options }: PiniaPluginContext<Id, S, G, A>) {
   // HMR module
   if (store.$id.startsWith('__hot:')) {
@@ -528,29 +564,23 @@ export function devtoolsPlugin<
   }
 
   // detect option api vs setup api
-  if (options.state) {
-    store._isOptionsAPI = true
-  }
+  store._isOptionsAPI = !!options.state
 
-  // only wrap actions in option-defined stores as this technique relies on
-  // wrapping the context of the action with a proxy
-  if (typeof options.state === 'function') {
+  patchActionForGrouping(
+    store as StoreGeneric,
+    Object.keys(options.actions),
+    store._isOptionsAPI
+  )
+
+  // Upgrade the HMR to also update the new actions
+  const originalHotUpdate = store._hotUpdate
+  toRaw(store)._hotUpdate = function (newStore) {
+    originalHotUpdate.apply(this, arguments as any)
     patchActionForGrouping(
-      // @ts-expect-error: can cast the store...
-      store,
-      Object.keys(options.actions)
+      store as StoreGeneric,
+      Object.keys(newStore._hmrPayload.actions),
+      !!store._isOptionsAPI
     )
-
-    const originalHotUpdate = store._hotUpdate
-
-    // Upgrade the HMR to also update the new actions
-    toRaw(store)._hotUpdate = function (newStore) {
-      originalHotUpdate.apply(this, arguments as any)
-      patchActionForGrouping(
-        store as StoreGeneric,
-        Object.keys(newStore._hmrPayload.actions)
-      )
-    }
   }
 
   addStoreToDevtools(