+++ packages/pinia/test-dts/store.test-d.ts
@@ -1,5 +1,5 @@
-import { StoreGeneric, defineStore, expectType } from './'
-import { watch } from 'vue'
+import { StoreGeneric, acceptHMRUpdate, defineStore, expectType } from './'
+import { UnwrapRef, watch } from 'vue'
 
 const useStore = defineStore({
   id: 'name',
@@ -80,6 +80,32 @@ defineStore({
   },
 })
 
+interface Model {
+  id: number
+}
+
+// Define generic factory function
+export function init<User extends Model>(name = 'settings') {
+  return defineStore(name, {
+    state: () => {
+      return {
+        // Set one of the properties to the generic type
+        user: {} as User,
+      }
+    },
+    actions: {
+      // Add action which accepts argument with our generic type
+      set(u: UnwrapRef<User>) {
+        // See linter error when trying to assign arg value to the state
+        this.user = u
+      },
+    },
+  })
+}
+
+const s = init()()
+s.set({ id: 1 })
+
 // getters on not existing properties
 defineStore({
   id: '',
@@ -115,6 +141,10 @@ defineStore({
 })
 
 const store = useStore()
+
+if (import.meta.hot) {
+  import.meta.hot.accept(acceptHMRUpdate(useStore, import.meta.hot))
+}
 
 expectType<{ a: 'on' | 'off' }>(store.$state)
 expectType<number>(store.nested.counter)